feat: return license and description metadata as part of build
Returning the meta.{description,license} fields from the build subsystem for
manifest builds for consistency with it being the source of the same fields
for NEF builds. This also opens up the opportunity for the build system to
modify and set defaults for these values, which is conceptually a task for
the build subsystem.