fix: always get description from build result
Rather than sometimes get the description from the lockfile, and
sometimes from the build result, now that we have it in the build
result, always get it from there.

Since they can be null for manifest builds, add a specific test for
those cases to ensure things are piped through correctly in that
scenario.

TODO: we could get rid of the `check_package_metadata` call all togther
since it's no longer doing all that much.  And refactor the logic for
the base catalog page to be used.