feat: add --no-auto-setup flag
In some cases you want to initialize an environment in a repository that
has an existing project in it, but you don't want Flox to prompt you
about suggested packages. This flag prevents the CLI from prompting you
when you know you don't want the suggested packages.

Reasons why you might want that:
- You want to install packages from a particular catalog rather than the
  base catalog.
- You disagree with the existing auto-setup-provided hooks.
- You just prefer to write things yourself.