feat: add --bare flag for minimal init
A common workflow for those of us that like small manifests is to
manually delete all of the manifest contents after running `flox init`.
This is just an extra work when all you want is an empty manifest. This
flag initializes the environment with the following manifest:

```
version = 1
```

This is the smallest legal manifest.