fix: buildenv empty outputs_to_install bug
Set outputs to install with list context rather than scalar context.

Manifest builds are now published with empty outputs_to_install. This
reaches previously unreached code that sets outputs_to_install to all
outputs. Currently it errors with:
```
❌ ERROR: Can't use string ("1") as an ARRAY ref while "strict refs" in use
```

This is because the current line
```
$package->{"outputs_to_install"} = keys %{$package->{"outputs"}};
```

is making an assignment in scalar context, which sets
$package->{"outputs_to_install"} to the number of keys in outputs
instead of the actual keys.