fix(generations): Unknown version suggest upgrade
Suggest that the user upgrades their version Flox if we encounter a
metadata version that is unknown by the current version.

This makes it clearer that the environment has been modified by a newer
version and what they need to do to resolve it, rather than suggesting
that it's been corrupted.

I've kept the original error as the first line because it contains the
found version which may be useful when debugging support requests.

Error handling further up the stack will take care of the trailing
newline so it's not included in the assertion string.