fix(commands): Complete path arguments
Fix shell completion for arguments that take files or directories as
paths. Previously they would give incorrect suggestions of an empty
string and the value given for `argument()`:

    % flox activate -d <TAB>
          path

Partial arguments wouldn't expand to anything:

    % flox activate -d ~/<TAB>

Assuming that you haven't enabled a fallback completer:

- https://zsh.sourceforge.io/FAQ/zshfaq04.html#l55

Now both work. Although there's an oddity when used with a combination
of `zsh`, `zstyle ':completion:*' group-name ''`, and `ShellComp::Dir`
which causes duplicates to be printed:

    % flox activate -d ~/demo/<TAB>
    -- file --
    bash/      composed/  default/   dep1/      dep2/      hello/     nef/       profile/   services/
    -- file --
    bash/      composed/  default/   dep1/      dep2/      hello/     nef/       profile/   services/

We could workaround that by only using `ShellComp::File`, which also
includes dirs and doesn't appear to be affected by the same problem, but
it seems to be relatively specific to my `zimfw` setup and doesn't
actually break usage.

It's also possible to restrict the masks for arguments that take `.toml`
or `.key` files but I don't think that filtering them provides a much
better experience and might cause confusion if testing with non-standard
file extensions.

I've mostly focused on `zsh` because that's what I use but I've also
tested this in `bash` with `flox install bash bash-completion` and
manually sourcing the completion stubs.