fix(ci): CodeCov from merge queue
Send test reports to CodeCov when running from the merge queue. It seems
this conditional may have been a mistake because we set it at the same
time as forcing both runs to have the same `branch_name`, so that the
results are collated together, but excluded the merge queue from runs.

Before that change:

- commit: https://github.com/flox/flox/commit/87b38f1fb9d7db77b5aee71003ba033a3b6b8792
- merge: https://github.com/flox/flox/actions/runs/16978182037/job/48132841669
- push: https://github.com/flox/flox/actions/runs/16978656531/job/48133848083

After that change, job skipped:

- commit: https://github.com/flox/flox/commit/bdc27284e52239c287f15c30ae8bf07991d85c25
- merge: https://github.com/flox/flox/actions/runs/16978824316/job/48134381261
- push: https://github.com/flox/flox/actions/runs/16979135480/job/48135251182