chore: drop dev integration tests
The dev tests are supposed to catch differences between a cargo built
`flox` and a Nix built `flox`, but they rarely catch anything so it
currently seems they aren't worth the resources.

We could continue to run the dev tests onto changes to *.nix files or
other files that may change the development environment, but that would
require some weird conditionals since we run the unit tests in the same
job as the dev integration tests, so I think it's simpler for now to
drop most of them.

I kept running cross-system.bats in the GH runners since it's not
trivial to pass secrets to runners.