fix(ci): CodeCov on all branches
So that we get flakey test reports from the merge queue.

I previously thought in 9f7756cab and 0abdc44f0 that filtering on the
`main` branch was enough to do this, based on the branch field in the
debug output of the CodeCov step, but the step was skipped when that PR
was in the merge queue:

- https://github.com/flox/flox/actions/runs/17619311246/job/50060919500

It's actually CodeCov that now converts the branch name from the
`gh-readonly-queue` prefixed name to the target branch name:

- https://github.com/getsentry/prevent-cli/commit/9f34ed801e03913eeea8cbf6f57399472a379f14

Sending the reports on all branches also means that we get the nice test
summary comments on PRs so that you can see which tests failed without
scrolling through the GitHub Actions output. We should have unlimited
uploads because we're both a) using a public repo, and b) are on the Pro
plan:

- https://about.codecov.io/pricing/
- https://app.codecov.io/plan/github/flox