fix(cuda): Detect more libnv*
There are some other libraries that could be matched by `libnv*.so*` in
addition to `libnvcuvid.so.*` from the preceding commit.

I can't personaly attest to what each of them do but it seems that they
are all NVIDIA related and may prevent CUDA detection from working in
some cases.