feat: Add build command to import from nixpkgs
Previously, to import a package definition from nixpkgs, there was quite the
incantation and it relied upon setting the EDITOR enviornment variable to 'cat'
which kind of screams hack. This change adds a `flox build import-nixpkgs
<pkg>` option to allow a one-line operation that imports the nixpkgs package
definition into .flox/pkgs/<pkg>.nix. If that file already exists, the command
will refuse and inform the user that they may choose to override that file with
a `--force` option.