test(activate): Profile script modifying PATH
Test the behaviour described in this bug report:

> If you have a default environment and a 2nd environment with:
>
> ```
> version = 1
>
> [profile]
> zsh = '''
  > export PATH="/foo:$PATH"
> '''
> ```
>
> Activating with zsh gives:
>
> ```
> > echo "$PATH"
> 2nd environment/bin:2nd environment/sbin:default/bin:default/sbin:/foo:...
> ```
>
> /foo should be at the start of PATH

This currently only works for bash, so we skip the tests for other
shells until they are fixed.

In an ideal world we would also assert that the project and default
environment PATHs are applied correctly but matching them in a regex
makes the test more fiddly and easier to get wrong.