fix(activate): zsh profile script can modify PATH
Fix the failing test for zsh with the suggestion:

> An alternative would be exporting a variable similar to what we already do with _flox_already_sourcing_bashrc and use that to so that activations in RC files only source profile scripts for FLOX_ENV. Using env variables to communicate between two activations seems error prone, but it’s been working okay for bash

I explored this other suggestion for a while, which would be preferable,
but identifying the number of places it *might* break which we might not
have test coverage for and debugging the recursion broke my head:

> Although for the first half of the proposal, the suggestion is not adding to FLOX_ENV_DIRS until after starting userShell. I don’t think we can do that without a decent amount of re-work, because etc-profiles, which runs before userShell, depends on FLOX_ENV_DIRS containing FLOX_ENV. Refactoring how we use FLOX_ENV_DIRS might end up being a code quality improvement but I’m unsure how difficult that would be to do.

These were known caveats of:

- 35a5eec5219b93943e905e60c9c17feda4426998
- https://flox-dev.slack.com/archives/C05P6A5J6U8/p1757427682740799

This approach doesn't currently work for fish or tcsh because they
automatically load RC files first, which sources profile scripts for all
environments and then re-orders the modified path after all
environments.

I briefly experimented with applying `--no-config` and the following to
fish but it seemed to still be double loading the config file in some of
our tests. Plus it breaks our support for fish 3.2 from 281f094:

    if [ -f ~/.config/fish/config.fish ] && [ -z "${_flox_sourcing_rc:=}" ]; then
      echo "set -gx _flox_sourcing_rc 1;"
      echo "source ~/.config/fish/config.fish;"
      echo "set -e _flox_sourcing_rc;"
    fi

This seems OK for now given:

- zsh and bash are our most popular shells
- zsh is already structured differently to the other shells
- current uncertainty about refactoring activation