fix(activate): Remove FLOX_ACTIVATION_PROFILE_ONLY
This was originally introduced for bash in fa66016 (plus other shells in
other commits) as a solution to `PATH` and `FPATH` being incorrectly
ordered when replaying activations for nested environments.

We've since fixed the ordering of `PATH`, `FPATH`, and several other
vars by re-constructing them correctly from `FLOX_ENV_DIRS` in
`flox-activations`.

So we should be able to remove this logic, which is hard to read and
threaded all the way through the interpreter.

This changes the behaviour of 8 related tests (2 groups of 4 shells)
because we now replay environments for the first in-place activation
when running a second in-place activation. This seems like the correct
behaviour to me and I'm not clear whether the existing tests were
intentional or documenting a side-effect.