feat(activate): Wire up --generation
Open a managed or remote environment at a specified generation,
including detecting the generation from an existing activation.

Uses new `RenderedEnvironmentLinks` so that the live generation
(including the separate links for `RemoteEnvironment`) remain
untouched. These links are not automatically GCed because we don't
currently have a way to recurse from the environment registry to check
in-use activations.

I originally implemented this with only the new links but other commands
that use the lockfile such as `flox list` would have incorrectly fallen
back to the live generation. By threading the "activated" generation
through to `ManageEnvironment` and thus `RemoteEnvironment` everything
subsequently uses the correct generation.

I also experimented with adding parallel constructor methods that took a
generation, rather than adding an optional generation everywhere, but it
was harder to read.