feat(activate): Record active --generation
Record when an environment is activated with `--generation` in
`_FLOX_ACTIVE_ENVIRONMENTS` which will be used in subsequent commits to
ensure that Flox commands within the activation continue to use the same
generation.

Includes a unit test for forward-compatibility from the existing format
of the environment variable. We can add other fields such as activation
mode (`run` or `dev`) to fix the bug where services don't work in the
default runmode environment.

I considered nesting the new fields under an `activation` field but it
doesn't seem necessary in terms of compatibility and slightly increases
the size of the env var which may become an issue one day.

I haven't changed the existing signatures or added an accessor methods
for `ActiveEnvironment` because we don't currently have a need for it
and pulling out the `.environment` for the current cases would be ugly.