feat(generations): Page output from list+history
The output from `generations list` and `generations history` can get
very long and spam the terminal for older environments that have lots of
generations.

The pager isn't invoked if the output would fit in the terminal, the
terminal isn't interactive, or the `--no-pager` argument is given.

Output has been reversed so that the most recent items appears at the
top of page. This behaviour doesn't change when not using a pager,
partly because we don't have control over when the pager decides to
page, but it's also consistent with how `git log` works.

This library seemed like a good choice and less fiddly than invoking an
external pager. There hasn't been a recent release but it looks pretty
feature complete. There are a few open issues but:

- I can't reproduce any problem with tracing
- text selection works for me when holding shift

The only thing that it's missing is the ability to persist the screen
when exiting the pager but that doesn't feel like a deal breaker when
you can disable the pager with an argument or by piping to `cat`.

I've verified the following features work fine for me:

- keyboard and mouse scrolling
- search, including incremental
- window resizing
- line wrapping
- metrics are sent when the pager is closed

There's not really much that we can write automated tests for, other
than getting output when non-interactive, but even that overlaps with
upstream tests and doesn't seem worth doing.

This doesn't appear to have affected the binary size by much:

    # Before
    % du -k cli/target/release/flox
    29148 cli/target/release/flox

    # After
    % du -k cli/target/release/flox
    29720 cli/target/release/flox