fix(activate): Omit message on activate -- command
We shouldn't print a message on `activate -- command` because it:

1. tells you to use `exit` to leave an activation which you don't have
   an interactive shell for
2. also tells you to use `exit` when using `flox/flox-direnv` which
   causes you to exit the entire shell rather than just the activation
3. clobbers STDERR from the command that's being executed

This was working correctly for "attach":

    flox [default term]
    (2) ~/demo/hello
    % flox activate
    ✅ You are now using the environment 'hello'.
    To stop using this environment, type 'exit'

    flox [hello default term]
    (3) ~/demo/hello
    % flox activate -- hello
    Hello, world!

But not for "start":

    (2) ~/demo/hello
    % flox activate -- hello
    ✅ You are now using the environment 'hello'.
    To stop using this environment, type 'exit'

    Hello, world!

Which we regressed when refactoring `start` into a function where `$#`
is always non-zero. Fix this by detecting the mode earlier on and
explicitly storing it in a local variable.

I haven't add any new tests because it's pretty fiddly to capture the
complete matrix of interactive x non-interactive x start x attach,
whilst also asserting on the complete output when we have a lot of noise
from the RC files and expect.