fix: improve switch to live generation error
We include a bunch of generic info that isn't actually relevant when
attempting to switch to the live generation

Also tweak the wording from "rollback to current" -> "switch to live"
since we shouldn't ever attempt to rollback to current (fixed in an
upcoming commit), and we use live instead of current now.

Before:
```
❌ ERROR: Failed to create a new generation: cannot rollback to current generation

This may be due to a corrupt environment
or another process modifying the environment.

Please try again later.

```

After:
```
❌ ERROR: Cannot switch to live generation
```