fix: drop extra set-env-dirs and fix-paths calls
As discussed in
https://github.com/flox/flox/pull/2552#discussion_r1924220017
these calls are defensive, but we don't actually have a clear case where
they're needed.

The behavior was preserved when porting _flox_env_helper to fix-paths
https://github.com/flox/flox/pull/2665/files#r1939927608,
but there wasn't any reason identified for the behavior in the first
place.

The only case I can think of this being relevant is:
1. User runs `flox activate`
2. User's .zshenv prepends a binary to PATH
3. We fix paths and put a FLOX_ENV that contains the binary at the front
   of path
4. User's .zshrc uses binary from PATH

Presumably the user would have written their .zshrc to use the binary
not provided by Flox, since that's what would happen when opening a new
shell.

So I don't think we need to re-prepend to PATH. Granted, that's not
entirely consistent with the fact that we prepend to PATH before running
the very first dotfile. But I'm not convinced that's best either.

Since there's not a motivating use case, I think we should opt for the
simpler behavior.