refactor(activate): pass vars explicitly for zsh
Instead of passing variables via environment variables, pass them with a
generated zsh startup script.

This is an improvement since:
1. We've had multiple activation bugs caused by environment variables
   being changed by nested activations
2. It paves the way to avoid leaking environment variables to the user

This should not change the behavior of what variables are exported at
this point since of the 7 variables changed,
- 4 are now exported explicitly in the zsh initialization script
- _activate_d, _FLOX_ACTIVATION_STATE_DIR, and _flox_activate_tracer are
  currently exported by assets/environment-interpreter/activate.

FLOX_ORIG_ZDOTDIR, ZDOTDIR, and FLOX_ZSH_INIT_SCRIPT are still passed by
exporting environment variables. We could avoid that by generating the
entirety of ZDOTDIR, but that would:
1. Require more runtime file generation which is slow
2. Be a more involved change

Since we've already ensured those variables won't be interfered with by
nested activations, I think that approach is stable enough to leave as
is for those variables.

I left _flox_activate_tracer because I think we want to drop exporting
that completely