chore: update test mock data
A few tests depend on rendered lockfiles to be unchanged.
As this PR updates the rendereing of lockfiles (even though minimally)
_some_ envs had to be updated.
Yet due to entenglement of resolution mocks and envs,
some resolution mocks had to be updated too.
Finding all the inidividual mocks to be updated (and become somewhat
inconsistent with the rest) seems unnecessary,
so this commit updates _all_ `envs` and `resolve` mocks.