fix: add build time env vars
Add variables to Nix builds depended on by the prior commit.

After this change, the CLI depends on both flox-activations and
flox-interpreter. It passes flox-interpreter as an argument to
flox-activations. flox-interpreter also depends on flox-activations

This is circular and could lead to two different flox-activations being
used, so in the future it would be better for flox-activations to be
self-contained such that the CLI would only have to depend on one thing.