fix: add custom schema implementation for `EnvironmentRef`
`EnvironmentRef` represents a _parsed string_, but the derived schema
it will be described as an object, which would result in parsing errors
of manifests generated strictly by the schema.
`EnvironmentRef` is only used in the context of remote includes,
so the impact is relatively small in case the previous schema
was already in use.