fix: exec commands directly from activate script
Our method of introducing a shell invocation between the `activate`
script and the commands to be invoked prevented signals from being
communicated from the pid of the `flox activate` invocation to the
resulting commands being invoked.

This patch adds conditional logic to replace that extra shell invocation
with a direct `exec` when running under containerd, so that the invoked
command can assume the PID of the `flox activate` invocation.