fix: reimplement jwt expiry validation
`jsonwebtoken` depricated (and completely removed the implementation of)
validating tokens without validating the signature in
<https://github.com/Keats/jsonwebtoken/pull/441>.

4f6dd581f41286b1a93502bd109a3f4db534c181 falsely assumed
that while deprecated,
`Validation::insecure_disable_signature_validation`
would still be effective.

While that is true, `flox auth login`
would now error with `InvalidKeyFormat`,
on account of attempting to parse the dummy value
we used to pass into the decoder (which was originally ignored)
as RS256 (advertised by auth0).
This issue was previously filed upstream as
<https://github.com/Keats/jsonwebtoken/issues/438>.
And is a behaviour change starting with the crypto refector of v10.x.

This change replaces `jsonwebtoken::decode` with
`jsonwebtoken::dangerous::insecure_decode`, which skips _all_ validation,
and manually reimplements expiry validation
on top of the raw unvalidated claims.

This is considered "safe" as the decoding only informs us of
1. the expriry date (for eager cli side management of the token)
2. the user handle for messaging and argument defaults

Auth and validation are only performed server side!