chore: remove watchdog, to extent possible
Where possible, removed references to the watchdog. In tests, replaced
"watchdog" with "executive" as a placeholder for reassessing the need
for those tests going forward. The watchdog was responsible for deleting
logs, and I expect some part of that functionality will need to be copied
back into the executive eventually.