ci: skip setting up nix for build job
Nix builds are run remotely through ssh,
thus we dont actually need a working nix installation
for the nix-build job.

This commit adds a flag to the common-setup to skip setting up nix,
and uses that to run `nix-build` without installing nix.

By way of skipping a nix installation, this also happens to save ~10s
off the setup's runtime.