fix(push): Logic error in push message
The logic for the `new` argument was inversed which caused us to always
say that environments were updated when there weren't any changes and
incorrectly stated that existing environments were always public.

On first glance I thought that we'd need a FloxHub API to determine
whether the environment was currently public or private but it's not
needed here because we only print the message when creating the
environment.

I'll create a separate feature request for creating a private
environment from the CLI which will require an API.

These new tests are a little bit slower than I'd like, presumably
because they have to interact with `git` in varying amounts, but they're
still faster than the BATS alternative and I think they're worth having
to prevent simple mistakes like this:

    PASS [   0.709s] flox::bin/flox commands::push::tests::push_new_environment
    PASS [   0.954s] flox::bin/flox commands::push::tests::push_no_changes
    PASS [   1.733s] flox::bin/flox commands::push::tests::push_changes