build: use `nix-unit` from nixpkgs
Upstream has merged our contribution for flake support long ago,
but didnt tag a release for a long time due to incompatibilities
with newer version of nix.
`nix-unit` has since been updated to compatibility with nix **2.30**,
and is now included in nixpkgs with a pinned reference to nix 2.30.
We can thus drop our own pin of `nix-unit`.