fix: add support for deprecated options in activate script
We have a test which expects to be able to invoke the activate script
from the interpreter of the current version of flox from a prior
version. That's like saying that we expect flox to be "future-compatible",
and there's a good reason that you don't hear that term as often as you
do "backwards-compatible".

So I'm not at all convinced of this "fix" but it makes the test pass
for now ...