fix: refactor generate_zsh_startup_commands to be like the others
Refactored generate_zsh_startup_commands() to be used like its
counterparts for the other shells, refactoring out a bunch of
common functionality in the process. This is a rough draft; we
need to more precisely rationalize what things need to be in
each of the zsh script and the zsh.rs file; it's just the first
version to bring the test failure count down.