refactor: `flox-activations activate` as entrypoint
Call `flox-activations activate` from the CLI rather than the Bash
activate script.

This should be functionally equivalent to current activation.

It does drop Sentry metrics for in-place activations. We're already
dropping them for command and interactive activations, and we can submit
them in the future in the executive if we want.