refactor: Consolidate custom package logic
The logic was duplicated in two places and tested in three places, which
made it hard to tell if they were all behaving the same.

Consolidate it onto a function in `manifest::raw` and the testing into
`CatalogPackage` since they are reponsible for parsing `pkg-path`.

I've left some basic tests in `manifest::typed` just to confirm that
`CatalogPackageDescriptor` delegates correctly.

I considered storing a `catalog: Option<String>` field on
`CatalogPackage` at parse time but felt like it might be a footgun when
`CatalogPackage.pkg_path` is still the canonical source.