feat: Remove slash prefixes from install IDs
Removes custom catalog prefixes from user-friendly install IDs because
they are unnecessary when:

- `flox edit` (and the manifest) already has the full `pkg-path`
- `flox list` now shows the full `pkg-path` after 8a7aac5ec

For example with `flox install flox/flox-mcp-server` this changes..

Manifest, before:

    "flox/flox-mcp-server".pkg-path = "flox/flox-mcp-server"

Manifest, after:

    flox-mcp-server.pkg-path = "flox/flox-mcp-server"

List, before:

    flox/flox-mcp-server: flox/flox-mcp-server (0.1.1)

List, after:

    flox-mcp-server: flox/flox-mcp-server (0.1.1)

This only affects new installs. We shouldn't attempt to modify existing
install IDs because it would break existing references. Uninstalls by
`pkg-path` continue to work as before.

This has a side-effect on package names that contain slashes regardless
of whether they are from a custom catalog, as demonstrated by the
changes to the unit tests, which I think is desirable. Duplicate install
IDs will still result in an error and can be resolved with `-i`.

We could be more clever about detecting duplicate install IDs that have
a different `pkg-path` but I don't think it's worth writing a ticket
until we get user feedback that it's a problem:

    % flox list
    hello: hello (2.12.2)
    % flox install -i hello cowsay
    ⚠️  Package with id 'hello' already installed to environment 'dcarley/tmp'