feat: layered verbosity for flox-activations
This introduces additional methods for setting the verbosity in
`flox-activations` with the following priority order:
- RUST_LOG
- _FLOX_ACTIVATIONS_VERBOSITY
- The existing `--verbosity` flag

The intention is that during development it's more convenient to type
out the flag than it is to type out the full environment variable. At
the same time you can't expect the shell scripts that call the
executable to juggle different values of the flag. And finally, when
debugging with a user, you may want to see output from sources other
than our own, so you need access to the full `RUST_LOG` variable.

The internal `_FLOX_ACTIVATIONS_VERBOSITY` variable is there to allow
us to set our own defaults for the various verbosity levels that give
useful information to the user, rather than requiring that they
understand how to use the variable themselves or simply giving them
all debug output from all sources.