feat(push): add --remote/-r flag for remote environments
Since remote environment no longer sync with FloxHub
with every interaction, we need a way to explicitly sync
remote environments.
This commit adds a `--remote/-r` flag
to identify the remote environment to push.
When pushing remote environments the same conflict semantics apply
as for managed environments.

Specifically, we factor out the existing `--dir/-d`
and `--owner/-o` flags into an enum `PushMode`
alongside the new remote flag.
The `handle` implementation matches on this to branch
either into the existing directory push logic or the new remote logic.

NOTE: with the semantics now closer to the remaining commands,
we could refactor this further to use `EnvironmentSelect` instead.