feat: add is_cached method to check without cloning
Add RemoteEnvironment::is_cached() to check if a remote environment exists
locally without triggering a clone from FloxHub.

This can be used in push and pull commands to verify
that the environment has previously been pulled,
i.e. that we actually have an "offline" state of the environment.
Checking this should be a rather cheap operation compared to
unnecessarily pulling the environment first
before pushing it back without changes.