feat: flox-activations containerize entrypoint
In the container case we build in some aspects of the activation
context, but some aspects must be determined at runtime since a
user can override the container's `Cmd` at runtime.

When no `Cmd` is provided, we set an "interactive" invocation
type. When arguments are provided either in the containerize
config or at runtime, we supply those as the `ActivateCtx.run_args` with
a "command" invocation type.

This also adds a new field `ActivateCtx.remove_after_reading` to inform
whether the file containing the `ActivateCtx` JSON should be deleted
before the `exec`. In the non-container case, where this is a tempfile,
that makes sense. In the container case, that JSON file is produced by
`mkContainer.nix` and placed in the store, which makes deleting it an
error.