fix: use build mode for version command
We don't currently use build mode when running build.*.version.command.
That means that if you have a `hook.on-activate`, it won't run for `flox
build` when you don't have a `version.command`, but if you do have a
`version.command`, `hook.on-activate` runs. That behavior is surprising,
so use `--mode build` instead to use the same method of activation as
other parts of build.