refactor: providers::services -> process_compose
Move the current services provider module to a sub-module named
`process_compose` so that we can co-locate a sibling module for
`systemd`.

The backends aren't strictly interchangeable and there isn't anything in
the `process_compose` module that's shareable by `systemd`. Even
`ServiceError` is very `process-compose` specific.

This will make it easier to navigate since the existing module was
already quite large, which a lot of standalone helper functions. And if
we decide not to proceed with persist services in the future then it
will make it easier to rip out the `systemd` specific code.