feat(systemd): Add new services provider
For persisting manifest defined services to systemd.

I've implemented basic overrides and merging where appropriate. For
examples, environment variables are always merged and systemd entries
take precedence.

We have to call `bash` in the exec commands because services commands can
be shell scripts and `process-compose` executes them in bash. I opted
for command activations rather than in-place activations because it
makes quoting significantly easier and it feels like a more natural use
of the feature. There are some improvements to quoting and multi-line
commands that will come in subsequent commits.

There aren't any unit tests for rendering since it's the responsibly of
our `systemd` crate.

There aren't any integration tests because I suspect they will be slow
and unreliable to co-ordinate, especially on non-ephemeral machines like
our builders where we can't guarantee clean state each time.

I've been testing manually with the following manifest:

    version = 1

    [install]
    hello.pkg-path = "hello"

    [services.hello]
    command = """
      while true; do
        hello
        sleep 2
      done
    """
    systemd.unit.description = "ello ello ello"

Which I did from macOS by using a Ubuntu 25.04 VM in Lima:

- https://lima-vm.io/docs/

I cheated a bit by modifying the OS check in `flox services persist` to
allow macOS and then copying the file into place in the VM, so that I
didn't have to build Flox in the VM:

    cp /Users/dcarley/.config/systemd/user/hello.service ~/.config/systemd/user/

This also required modifying the VM template to mount `~` as writable so
that Flox could build the environment for Linux when the unit starts.

Then I reloaded the unit with:

    systemctl daemon-reload --user && systemctl restart hello --user

In order to get any journal logs for user services you'll also need to
modify the UID in the template because by default it will map from your
macOS user which was 501 for me and causes the logs to be collected by
the system journal..

From: https://wiki.archlinux.org/title/Systemd/User

> journald will not write user journals for users with UIDs below 1000,
> instead directing everything to the system journal.

From: https://systemd.io/UIDS-GIDS/

> 1…999 → System users. These are users that do not map to actual
> “human” users, but are used as security identities for system daemons,
> to implement privilege separation and run system daemons with minimal
> privileges.
>
> 1000…65533 and 65536…4294967294 → Everything else, i.e.
> regular (human) users.

There's also *another* bug with users services that will be fixed in a
subsequent commit.