feat(services): Add persist command for systemd
This currently writes files with placeholder contents until the provider
is wired together.

There are some TODOs that we don't need to implement for the first cut.

I've used `ServicesEnvironment` for consistency with the other commands
but it's not actually doing much now that we support services for all
types of environment so I may follow-up with a PR to remove it.