fix(services): Disable .env support
Prevent `process-compose` from loading `.env` files because it causes
environment variables to shadow `[vars]` and `[services.*.vars]` in the
manifest:

- https://f1bonacc1.github.io/process-compose/configuration/#env-file

This functionality was added upstream in:

- https://github.com/F1bonacc1/process-compose/releases/tag/v1.24.0

Users that want `.env` support should source the file from
`hook.on-activate` instead so that it's layered correctly.

Prior to the code change the new test fails with:

    -- output does not contain substring --
    substring (1 lines):
      right_value
    output (2 lines):
      ✅ Service 'one' started.
      hello wrong_value
    --

The duplicate `process-compose up` lines could benefit from being
refactored into conditional args in an array but I believe they're going
to change in the ongoing activate refactor.