fix: only print some exports for in-place
The CLI used to print out all exports added to the Command created for
running the activate script.
During the refactor, I accidentally added additional export statements
to the in-place command by adding env vars to the command used to run the
activate script command. Split out just the variables that the CLI was
previously adding