feat(systemd): Support remote environments
Calling `flox services persist` for a remote environment (either
activated imperatively or passed as a declarative argument) now renders
a unit file that calls `flox activate -r` rather than referring to the
cached managed environment.

This ensures that the unit automatically gets updates when it starts or
restarts and it's clearer for the user who shouldn't really see the
cached managed environment.

I deliberated what to call and where to put `ActivateEnvironmentRef` and
the argument rendering function. It's kinda analogous to the CLI's
`EnvironmentSelect` but the environment has already been resolved. We
only have one use of it but it felt too generic to hide under the
`systemd` provider.

This is another place that might benefit from being able to pin the
generation of a managed or remote environment but we don't currently
have a consistent way of surfacing that to users and we don't know
whether the persist feature will be very popular, so I'm not going to
add it at the moment.

I have not manually tested this in a Linux VM because I'm 99% confident
based on the unit tests and behaviour of `flox activate`.